requirements.txt
Flask
python-dotenv
flask-cors
google-generativeai
